create component for input form
