+++ src/App.jsx
@@ -1,22 +1,10 @@
 import { useState } from "react"
+import { NewTodoForm } from "./NewTodoForm"
 import "./styles.css"
 
 export default function App(){
-  const [newItem, setNewItem] = useState("")
+
   const [todos, setTodos] = useState([])
-
-  function handleSubmit(event) {
-    event.preventDefault()
-
-    setTodos(currentTodos => {
-      return [
-        ...currentTodos,
-        {id: crypto.randomUUID(), title: newItem, completed: false},
-      ]
-    })
-
-    setNewItem("")
-  }
 
   function toggleTodo(id, completed) {
     setTodos(currentTodos => {
@@ -37,18 +25,7 @@ export default function App(){
 
   return (
     <>
-      <form onSubmit={handleSubmit} className="new-item-form">
-        <div className="form-row">
-          <label htmlFor="item">New Item</label>
-          <input 
-          value={newItem} 
-          onChange={event => setNewItem(event.target.value)} 
-          type="text" 
-          id="item" 
-          />
-        </div>
-        <button className="btn">Add</button>
-      </form>
+    <NewTodoForm />
       <h1 className="header">Jamie's Todo List</h1>
       <ul className="list">
         {todos.length === 0 && "No Todos"}
